Use TLS when required for muxsrv
... instead of using TLS just for the ReST server.
